+++ reana_commons/validation/utils.py
@@ -1,17 +1,49 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of REANA.
-# Copyright (C) 2021, 2022 CERN.
+# Copyright (C) 2022 CERN.
 #
 # REANA is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
-"""REANA-Commons workspaces util."""
 
+"""REANA-Commons validation utilities."""
+
+import json
+import logging
 import os
-from typing import List
+from typing import Dict, List
 
-from reana_commons.config import REANA_WORKFLOW_NAME_ILLEGAL_CHARACTERS, WORKSPACE_PATHS
+from jsonschema import ValidationError, validate
+
+from reana_commons.config import (
+    REANA_WORKFLOW_NAME_ILLEGAL_CHARACTERS,
+    WORKSPACE_PATHS,
+    reana_yaml_schema_file_path,
+)
 from reana_commons.errors import REANAValidationError
+
+
+def validate_reana_yaml(reana_yaml: Dict) -> None:
+    """Validate REANA specification file according to jsonschema.
+
+    :param reana_yaml: Dictionary which represents REANA specification file.
+    :raises ValidationError: Given REANA spec file does not validate against
+        REANA specification schema.
+    """
+    try:
+        with open(reana_yaml_schema_file_path, "r") as f:
+            reana_yaml_schema = json.loads(f.read())
+            validate(reana_yaml, reana_yaml_schema)
+    except IOError as e:
+        logging.info(
+            "Something went wrong when reading REANA validation schema from "
+            "{filepath} : \n"
+            "{error}".format(filepath=reana_yaml_schema_file_path, error=e.strerror)
+        )
+        raise e
+    except ValidationError as e:
+        logging.info("Invalid REANA specification: {error}".format(error=e.message))
+        raise e
 
 
 def validate_workflow_name(workflow_name: str) -> str:
@@ -45,7 +77,7 @@ def validate_workspace(
         )
         if not available:
             raise REANAValidationError(
-                f'Desired workspace "{workspace_option}" not valid.\n'
-                "Please run `reana-client info` to see the list of allowed prefix values."
+                f'Desired workspace "{workspace_option}" is not valid.\n'
+                f'Available workspace prefix values are: {", ".join(available_paths)}',
             )
     return workspace_option